feat(*) add database, schema, basic api and client
This commit introduces Embedded Derby as dbms, and a database schema
focused on "Media". The basic api allows only to read, and the client
webapp can display the data as a list, with details available when clicking an
item.
The database is currently dropped when the server stops, and the data is
populated on start from `server/src/main/resources/import.sql`